+++ src/main/java/com/example/service/BrandUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Brand;
 
 public interface BrandUpdateService {
-	Brand updateData(Brand brand , int id);
+	Brand updateData(Brand brand , Long id);
 
 }
