+++ src/main/java/com/example/service/CODFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.COD;
 
 public interface CODFetchService {
-	COD fetchById(int id);
+	COD fetchById(Long id);
 
 }
