+++ src/main/java/com/example/service/CODUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.COD;
 
 public interface CODUpdateService {
-	COD updateData(COD cod , int id);
+	COD updateData(COD cod , Long id);
 
 }
