+++ src/main/java/com/example/service/CardFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Card;
 
 public interface CardFetchService {
-	Card fetchById(int id);
+	Card fetchById(Long id);
 
 }
