+++ src/main/java/com/example/service/CardUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Card;
 
 public interface CardUpdateService {
-	Card updateData(Card card , int id);
+	Card updateData(Card card , Long id);
 
 }
