+++ src/main/java/com/example/service/CompanyResponseUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.CompanyResponse;
 
 public interface CompanyResponseUpdateService {
-	CompanyResponse updateData(CompanyResponse companyResponse , int id);
+	CompanyResponse updateData(CompanyResponse companyResponse , Long id);
 
 }
