+++ src/main/java/com/example/service/FeedbackUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Feedback;
 
 public interface FeedbackUpdateService {
-	Feedback updateData(Feedback feedback , int id);
+	Feedback updateData(Feedback feedback , Long id);
 
 }
