+++ src/main/java/com/example/service/InvoiceFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Invoice;
 
 public interface InvoiceFetchService {
-	Invoice fetchById(int id);
+	Invoice fetchById(Long id);
 
 }
