+++ src/main/java/com/example/service/InvoiceUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Invoice;
 
 public interface InvoiceUpdateService {
-	Invoice updateData(Invoice invoice , int id);
+	Invoice updateData(Invoice invoice , Long id);
 
 }
