+++ src/main/java/com/example/service/OrderFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Order;
 
 public interface OrderFetchService {
-	Order fetchById(int id);
+	Order fetchById(Long id);
 
 }
