+++ src/main/java/com/example/service/ProductReviewUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.ProductReview;
 
 public interface ProductReviewUpdateService {
-	ProductReview updateData(ProductReview productReview , int id);
+	ProductReview updateData(ProductReview productReview , Long id);
 
 }
