+++ src/main/java/com/example/service/ProductUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Product;
 
 public interface ProductUpdateService {
-	Product updateData(Product product , int id);
+	Product updateData(Product product , Long id);
 
 }
