+++ src/main/java/com/example/service/ShippingDetailsUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.ShippingDetails;
 
 public interface ShippingDetailsUpdateService {
-	ShippingDetails updateData(ShippingDetails shippingDetails , int id);
+	ShippingDetails updateData(ShippingDetails shippingDetails , Long id);
 
 }
