+++ src/main/java/com/example/service/SubCategoryFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.SubCategory;
 
 public interface SubCategoryFetchService {
-	SubCategory fetchById(int id);
+	SubCategory fetchById(Long id);
 
 }
