+++ src/main/java/com/example/service/SubCategoryUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.SubCategory;
 
 public interface SubCategoryUpdateService {
-	SubCategory updateData(SubCategory subCategory , int id);
+	SubCategory updateData(SubCategory subCategory , Long id);
 
 }
