+++ src/main/java/com/example/service/UpiFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Upi;
 
 public interface UpiFetchService {
-	Upi fetchById(int id);
+	Upi fetchById(Long id);
 
 }
